requirements.txt
python-dotenv
flask
boto3
pathlib
requests
git+git://github.com/NYCPlanning/cook.git@master#egg=cook
